feat: add create and update for user

+++ src/user/user.service.ts
@@ -1,6 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { UpdatePasswordDto } from './dto/update-user-password.dto';
 import { InMemoryStorage } from 'src/storage/in-memory-storage';
 
 @Injectable()
@@ -8,27 +12,36 @@ export class UserService {
   constructor(private storage: InMemoryStorage) {}
 
   create(createUserDto: CreateUserDto) {
-    console.log('createUserDto-->', createUserDto);
-    return 'This action adds a new user';
+    return this.storage.createUser(createUserDto);
   }
 
   findAll() {
-    // return `This action returns all user`;
     return this.storage.getUsers();
   }
 
   findOne(id: string) {
-    // return `This action returns a #${id} user`;
     return this.storage.getUserById(id);
   }
 
-  update(id: string, updateUserDto: UpdateUserDto) {
-    console.log('updateUserDto-->', updateUserDto);
-    return `This action updates a #${id} user`;
+  update(id: string, updatePasswordDto: UpdatePasswordDto) {
+    const userForUpdate = this.storage.getUserById(id);
+
+    if (!userForUpdate) {
+      throw new NotFoundException('User not found');
+    }
+
+    if (updatePasswordDto.oldPassword !== userForUpdate.password) {
+      throw new ForbiddenException('Old password is wrong');
+    }
+
+    return this.storage.updateUser(id, updatePasswordDto);
   }
 
   remove(id: string) {
-    // return `This action removes a #${id} user`;
+    const res = this.storage.deleteUser(id);
+    if (!res) {
+      throw new NotFoundException('User not found');
+    }
     return this.storage.deleteUser(id);
   }
 }
